requirements.txt
pathspec>=0.11.0
PyYAML>=6.0
colorama>=0.4.6
